Removed the async_connect_waiting and blocking states from LmConnection.
As a first step of cleaning up the network code underlying LmConnection I want to make the blocking and non-blocking calls be as similar as possible.

The socket should be able to connect in asynchronous manner even if we want to block on the open call since the mainloop iteration will handle the "block" for lm_connection_open_and_block.