Only check for IO Condition before trying another read when the socket is blocking [#23]
In commit 79deb183d we added a check to see if there was more data to be read before trying again. This
to avoid having a blocking socket hang on that call until some data became available.

However, when using OpenSSL it sometimes reads everything off the TCP socket and buffers it internally
which means that the TCP socket will tell us that there is no more data available which leads to us not
trying to read from the OpenSSL layer again.

In this commit I added a check to see if the socket was blocking before doing this. The long term fix
will be that the socket is always going to be in async mode and the synchronous APIs will be handled at
a higher level.

[#23 responsible:Hallski milestone:Loudmouth 1.4.3 state:resolved]